LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_removed_interleaved.tentative_pen-expected.txt

child
done

FAIL mouse events from pen received before/after child removal at pointerdown assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mouseover@parent,mousedown@parent,mouseup@parent,click@parent,mousedown@parent,mouseup@parent,click@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mousedown@child,mouseup@parent,mouseout@child,mouseleave@child,mouseover@parent,mouseenter@parent,mouseout@parent,mouseleave@parent"
FAIL mouse events from pen received before/after child removal at pointerup assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseover@parent,mouseup@parent,click@parent,mousedown@parent,mouseup@parent,click@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseup@child,mouseout@child,mouseleave@child,mouseover@parent,mouseenter@parent,mouseout@parent,mouseleave@parent"
